fixed bars - color - mood - calculation of fat% and muscle% is off

+++ app/src/main/java/com/progressor/progressor/model/dataobjects/account/Body.kt
@@ -2,8 +2,8 @@ package com.progressor.progressor.model.dataobjects.account
 
 data class Body(
         val weight: Double? = null,
-        val fatPercentage: Double? = null,
-        val musclePercentage: Double? = null,
+        val fatPercentage: Int? = null,
+        val musclePercentage: Int? = null,
         val mood: Int = 0,
         val createDate: String = ""
 )
